move ChatMessage to its own component, wrapped by Chat
